Crate rusoto_autoscaling_plans

AWS Auto Scaling
Use AWS Auto Scaling to quickly discover all the scalable AWS resources for your application and configure dynamic scaling and predictive scaling for your resources using scaling plans. Use this service in conjunction with the Amazon EC2 Auto Scaling, Application Auto Scaling, Amazon CloudWatch, and AWS CloudFormation services.

Currently, predictive scaling is only available for Amazon EC2 Auto Scaling groups.

For more information about AWS Auto Scaling, including information about granting IAM users required permissions for AWS Auto Scaling actions, see the AWS Auto Scaling User Guide.

If you're using the service, you're probably looking for AutoscalingPlansClient and AutoscalingPlans.

Structs

ApplicationSource
Represents an application source.

AutoscalingPlansClient
A client for the AWS Auto Scaling Plans API.

CreateScalingPlanRequest

CreateScalingPlanResponse

CustomizedLoadMetricSpecification
Represents a CloudWatch metric of your choosing that can be used for predictive scaling.

For predictive scaling to work with a customized load metric specification, AWS Auto Scaling needs access to the Sum and Average statistics that CloudWatch computes from metric data. Statistics are calculations used to aggregate data over specified time periods.

When you choose a load metric, make sure that the required Sum and Average statistics for your metric are available in CloudWatch and that they provide relevant data for predictive scaling. The Sum statistic must represent the total load on the resource, and the Average statistic must represent the average load per capacity unit of the resource. For example, there is a metric that counts the number of requests processed by your Auto Scaling group. If the Sum statistic represents the total request count processed by the group, then the Average statistic for the specified metric must represent the average request count processed by each instance of the group.

For information about terminology, available metrics, or how to publish new metrics, see Amazon CloudWatch Concepts in the Amazon CloudWatch User Guide.

CustomizedScalingMetricSpecification
Represents a CloudWatch metric of your choosing that can be used for dynamic scaling as part of a target tracking scaling policy.

To create your customized scaling metric specification:

Add values for each required parameter from CloudWatch. You can use an existing metric, or a new metric that you create. To use your own metric, you must first publish the metric to CloudWatch. For more information, see Publish Custom Metrics in the Amazon CloudWatch User Guide.

Choose a metric that changes proportionally with capacity. The value of the metric should increase or decrease in inverse proportion to the number of capacity units. That is, the value of the metric should decrease when capacity increases.

For more information about CloudWatch, see Amazon CloudWatch Concepts.

Datapoint
Represents a single value in the forecast data used for predictive scaling.

DeleteScalingPlanRequest

DeleteScalingPlanResponse

DescribeScalingPlanResourcesRequest

DescribeScalingPlanResourcesResponse

DescribeScalingPlansRequest

DescribeScalingPlansResponse

GetScalingPlanResourceForecastDataRequest

GetScalingPlanResourceForecastDataResponse

MetricDimension
Represents a dimension for a customized metric.

PredefinedLoadMetricSpecification
Represents a predefined metric that can be used for predictive scaling.

PredefinedScalingMetricSpecification
Represents a predefined metric that can be used for dynamic scaling as part of a target tracking scaling policy.

ScalingInstruction
Describes a scaling instruction for a scalable resource.

The scaling instruction is used in combination with a scaling plan, which is a set of instructions for configuring dynamic scaling and predictive scaling for the scalable resources in your application. Each scaling instruction applies to one resource.

AWS Auto Scaling creates target tracking scaling policies based on the scaling instructions. Target tracking scaling policies adjust the capacity of your scalable resource as required to maintain resource utilization at the target value that you specified.

AWS Auto Scaling also configures predictive scaling for your Amazon EC2 Auto Scaling groups using a subset of parameters, including the load metric, the scaling metric, the target value for the scaling metric, the predictive scaling mode (forecast and scale or forecast only), and the desired behavior when the forecast capacity exceeds the maximum capacity of the resource. With predictive scaling, AWS Auto Scaling generates forecasts with traffic predictions for the two days ahead and schedules scaling actions that proactively add and remove resource capacity to match the forecast.

We recommend waiting a minimum of 24 hours after creating an Auto Scaling group to configure predictive scaling. At minimum, there must be 24 hours of historical data to generate a forecast.

For more information, see Getting Started with AWS Auto Scaling.

ScalingPlan
Represents a scaling plan.

ScalingPlanResource
Represents a scalable resource.

ScalingPolicy
Represents a scaling policy.

TagFilter
Represents a tag.

TargetTrackingConfiguration
Describes a target tracking configuration to use with AWS Auto Scaling. Used with ScalingInstruction and ScalingPolicy.

UpdateScalingPlanRequest

UpdateScalingPlanResponse

Enums

CreateScalingPlanError
Errors returned by CreateScalingPlan

DeleteScalingPlanError
Errors returned by DeleteScalingPlan

DescribeScalingPlanResourcesError
Errors returned by DescribeScalingPlanResources

DescribeScalingPlansError
Errors returned by DescribeScalingPlans

GetScalingPlanResourceForecastDataError
Errors returned by GetScalingPlanResourceForecastData

UpdateScalingPlanError
Errors returned by UpdateScalingPlan

Traits

AutoscalingPlans
Trait representing the capabilities of the AWS Auto Scaling Plans API. AWS Auto Scaling Plans clients implement this trait.