Struct rusoto_autoscaling_plans::AutoscalingPlansClient

pub struct AutoscalingPlansClient { /* fields omitted */ }

A client for the AWS Auto Scaling Plans API.

Methods

impl AutoscalingPlansClient

pub fn new(region: Region) -> AutoscalingPlansClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> AutoscalingPlansClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,