Trait rusoto_autoscaling_plans::AutoscalingPlans

pub trait AutoscalingPlans {
    fn create_scaling_plan(
        &self, 
        input: CreateScalingPlanRequest
    ) -> RusotoFuture<CreateScalingPlanResponse, CreateScalingPlanError>;

    fn delete_scaling_plan(
        &self, 
        input: DeleteScalingPlanRequest
    ) -> RusotoFuture<DeleteScalingPlanResponse, DeleteScalingPlanError>;

    fn describe_scaling_plan_resources(
        &self, 
        input: DescribeScalingPlanResourcesRequest
    ) -> RusotoFuture<DescribeScalingPlanResourcesResponse, DescribeScalingPlanResourcesError>;

    fn describe_scaling_plans(
        &self, 
        input: DescribeScalingPlansRequest
    ) -> RusotoFuture<DescribeScalingPlansResponse, DescribeScalingPlansError>;

    fn get_scaling_plan_resource_forecast_data(
        &self, 
        input: GetScalingPlanResourceForecastDataRequest
    ) -> RusotoFuture<GetScalingPlanResourceForecastDataResponse, GetScalingPlanResourceForecastDataError>;

    fn update_scaling_plan(
        &self, 
        input: UpdateScalingPlanRequest
    ) -> RusotoFuture<UpdateScalingPlanResponse, UpdateScalingPlanError>;
}

Trait representing the capabilities of the AWS Auto Scaling Plans API. AWS Auto Scaling Plans clients implement this trait.

Required methods

fn create_scaling_plan(
&self,
input: CreateScalingPlanRequest
) -> RusotoFuture<CreateScalingPlanResponse, CreateScalingPlanError>

Creates a scaling plan.

fn delete_scaling_plan(
&self,
input: DeleteScalingPlanRequest
) -> RusotoFuture<DeleteScalingPlanResponse, DeleteScalingPlanError>

Deletes the specified scaling plan.

Deleting a scaling plan deletes the underlying ScalingInstruction for all of the scalable resources that are covered by the plan.

If the plan has launched resources or has scaling activities in progress, you must delete those resources separately.

fn describe_scaling_plan_resources(
&self,
input: DescribeScalingPlanResourcesRequest
) -> RusotoFuture<DescribeScalingPlanResourcesResponse, DescribeScalingPlanResourcesError>

Describes the scalable resources in the specified scaling plan.

fn describe_scaling_plans(
&self,
input: DescribeScalingPlansRequest
) -> RusotoFuture<DescribeScalingPlansResponse, DescribeScalingPlansError>

Describes one or more of your scaling plans.

fn get_scaling_plan_resource_forecast_data(
&self,
input: GetScalingPlanResourceForecastDataRequest
) -> RusotoFuture<GetScalingPlanResourceForecastDataResponse, GetScalingPlanResourceForecastDataError>

Retrieves the forecast data for a scalable resource.

Capacity forecasts are represented as predicted values, or data points, that are calculated using historical data points from a specified CloudWatch load metric. Data points are available for up to 56 days.

fn update_scaling_plan(
&self,
input: UpdateScalingPlanRequest
) -> RusotoFuture<UpdateScalingPlanResponse, UpdateScalingPlanError>

Updates the specified scaling plan.

You cannot update a scaling plan if it is in the process of being created, updated, or deleted.

Loading content...