Crate rusoto_autoscaling_plans

AWS Auto Scaling
Use AWS Auto Scaling to quickly discover all the scalable AWS resources for your application and configure dynamic scaling for your scalable resources.

To get started, create a scaling plan with a set of instructions used to configure dynamic scaling for the scalable resources in your application. AWS Auto Scaling creates target tracking scaling policies for the scalable resources in your scaling plan. Target tracking scaling policies adjust the capacity of your scalable resource as required to maintain resource utilization at the target value that you specified.

If you're using the service, you're probably looking for AutoscalingPlansClient and AutoscalingPlans.

Structs

ApplicationSource
Represents an application source.

AutoscalingPlansClient
A client for the AWS Auto Scaling Plans API.

CreateScalingPlanRequest

CreateScalingPlanResponse

CustomizedScalingMetricSpecification
Represents a customized metric for a target tracking policy.

DeleteScalingPlanRequest

DeleteScalingPlanResponse

DescribeScalingPlanResourcesRequest

DescribeScalingPlanResourcesResponse

DescribeScalingPlansRequest

DescribeScalingPlansResponse

MetricDimension
Represents a dimension for a customized metric.

PredefinedScalingMetricSpecification
Represents a predefined metric for a target tracking policy.

ScalingInstruction
Specifies the scaling configuration for a scalable resource.

ScalingPlan
Represents a scaling plan.

ScalingPlanResource
Represents a scalable resource.

ScalingPolicy
Represents a scaling policy.

TagFilter
Represents a tag.

TargetTrackingConfiguration
Represents a target tracking scaling policy.

UpdateScalingPlanRequest

UpdateScalingPlanResponse

Enums

CreateScalingPlanError
Errors returned by CreateScalingPlan

DeleteScalingPlanError
Errors returned by DeleteScalingPlan

DescribeScalingPlanResourcesError
Errors returned by DescribeScalingPlanResources

DescribeScalingPlansError
Errors returned by DescribeScalingPlans

UpdateScalingPlanError
Errors returned by UpdateScalingPlan

Traits

AutoscalingPlans
Trait representing the capabilities of the AWS Auto Scaling Plans API. AWS Auto Scaling Plans clients implement this trait.